Bert overhaul using config and parse files overhaul done. Remaining is overhaul of dataloader classes and setup and train of bert model as well as check if prev files run with new config (#9)
